vehicle-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "vehicle-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "nats-py>=2.9.0",
    "pydantic>=2.7.0",
]

[tool.setuptools.packages.find]
where = ["."]
